[REFACTOR] : 전역 에러 처리기 보완
